Add support for zero-copy conversions from an xarray to a DimArray (#972)
* Add support for zero-copy conversions from an xarray to a DimArray

* Add precompilation statements for the PythonCall extension

* Attempt to fix OpenSSL compat issues in CI

If PythonCall installed a version of OpenSSL incompatible with what Julia is
built with and it loads the newer version first, that will conflict with any
other Julia package (e.g Plots.jl) that tries to load OpenSSL afterwards. Here
we hack around that by just loading OpenSSL.jl first so that Python uses that.